Add a Doc class and use it
