+++ ietf_reviewtool/metadata.py
@@ -7,42 +7,40 @@ import num2words
 
 from .review import IetfReview
 from .util.fetch import fetch_meta
-from .util.docposition import SECTION_PATTERN
-from .util.text import (
-    unfold,
-    word_join,
-    get_status,
-    get_relationships,
-)
+from .doc import Doc
+from .util.text import word_join
 
 
 def check_meta(
-    datatracker: str, text: str, meta: dict, review: IetfReview, log: logging.Logger
+    doc: Doc, review: IetfReview, datatracker: str, log: logging.Logger
 ) -> None:
     """
     Check document metadata for issues.
 
-    @param      text   The text of the document
-    @param      meta   The metadata
-    @param      review The IETF review to comment upon
+    @param      doc          The document
+    @param      review       The IETF review to comment upon
+    @param      datatracker  The datatracker
+    @param      log          The log
+
+    @return     { description_of_the_return_value }
     """
 
-    level = meta["std_level"] or meta["intended_std_level"]
+    level = doc.meta["std_level"] or doc.meta["intended_std_level"]
     if not level:
         review.discuss(
             "Datatracker does not record an intended RFC status for this document.",
         )
     else:
-        status = get_status(text)
-        if status != level and (
-            level != "Proposed Standard" or status != "Standards Track"
+        if doc.status.lower() != level.lower() and (
+            level.lower() != "proposed standard"
+            or doc.status.lower() != "standards track"
         ):
             review.discuss(
                 f'Intended RFC status in datatracker is "{level}", but '
-                f'document says "{status}".',
+                f'document says "{doc.status}".',
             )
 
-    num_authors = len(meta["authors"])
+    num_authors = len(doc.meta["authors"])
     if num_authors > 5:
         review.comment(
             f"The document has {num2words.num2words(num_authors)} "
@@ -52,7 +50,7 @@ def check_meta(
         )
 
     iana_review_state = (
-        meta["iana_review_state"] if "iana_review_state" in meta else None
+        doc.meta["iana_review_state"] if "iana_review_state" in doc.meta else None
     )
     if iana_review_state:
         if re.match(r".*Not\s+OK", iana_review_state, flags=re.IGNORECASE):
@@ -64,43 +62,44 @@ def check_meta(
                 "The IANA review of this document seems to not have " "concluded yet.",
             )
 
-    consensus = meta["consensus"] if "consensus" in meta else None
+    consensus = doc.meta["consensus"] if "consensus" in doc.meta else None
     if consensus is None:
         review.comment(
             "The datatracker state does not indicate whether the "
             "consensus boilerplate should be included in this document.",
         )
 
-    stream = meta["stream"] if "stream" in meta else None
+    stream = doc.meta["stream"] if "stream" in doc.meta else None
     if stream != "IETF":
         review.comment(
             "This does not seem to be an IETF-stream document.",
         )
 
-    status = get_status(text)
-    for rel, docs in get_relationships(text).items():
+    for rel, rel_docs in doc.relationships.items():
         if rel == "updates":
-            abstract = unfold(extract_abstract(text))
             missing_docs = []
-            for doc in docs:
-                if not re.search(r"RFC\s*" + doc, abstract):
-                    missing_docs.append(doc)
+            for rel_doc in rel_docs:
+                if not re.search(r"RFC\s*" + rel_doc, doc.abstract):
+                    missing_docs.append(rel_doc)
             if missing_docs:
-                updates = word_join(docs, prefix="RFC")
+                updates = word_join(rel_docs, prefix="RFC")
                 review.discuss(
                     f"This document updates {updates}, but does not seem "
                     f"to include explanatory text about this in the "
                     f"abstract.",
                 )
 
-        for doc in docs:
-            meta = fetch_meta(datatracker, "rfc" + doc, log)
+        for rel_doc in rel_docs:
+            doc.meta = fetch_meta(datatracker, "rfc" + rel_doc, log)
             level = (
-                meta["std_level"] or meta["intended_std_level"] if meta else "Unknown"
+                doc.meta["std_level"] or doc.meta["intended_std_level"]
+                if doc.meta
+                else "Unknown"
             )
-            if not relationship_ok(status, level):
+            if not relationship_ok(doc.status, level):
                 review.discuss(
-                    f"This {status} document {rel} RFC{doc}, " f"which is {level}.",
+                    f"This {doc.status} document {rel} RFC{rel_doc}, "
+                    f"which is {level}.",
                 )
 
 
@@ -122,27 +121,3 @@ def relationship_ok(status: str, level: str) -> bool:
         "internet standard",
     ]
     return (status.lower() in std) or (level.lower() not in std)
-
-
-def extract_abstract(text: list) -> str:
-    """
-    Return that abstract of the text .
-
-    @param      text  The text to parse for the abstract
-
-    @return     The abstract.
-    """
-    in_abstract = False
-    abstract = ""
-    for line in text.splitlines(keepends=True):
-        pot_sec = SECTION_PATTERN.search(line)
-        if pot_sec:
-            which = pot_sec.group(0)
-            if re.search(r"^Abstract", which):
-                in_abstract = True
-                continue
-            if abstract:
-                break
-        if in_abstract:
-            abstract += line
-    return abstract
