+++ ietf_reviewtool/util/text.py
@@ -224,76 +224,6 @@ def strip_pagination(text: str) -> str:
     return stripped
 
 
-def extract_refs(text: list) -> dict:
-    """
-    Return a dict of references found in the text as well as the normative and
-    informative reference sections.
-
-    @param      text  The text to parse for references
-
-    @return     A dict with sets of found references.
-    """
-    parts = {"text": "", "informative": "", "normative": ""}
-    part = "text"
-    for line in text.splitlines(keepends=True):
-        pot_sec = SECTION_PATTERN.search(line)
-        if pot_sec:
-            which = pot_sec.group(0)
-            if re.search(
-                r"^(?:(\d\.?)+\s+)?(?:Non-Norm|Inform)ative\s+References?\s*$",
-                which,
-                flags=re.IGNORECASE,
-            ):
-                part = "informative"
-            elif re.search(
-                r"^(?:(\d\.?)+\s+)?(Normative\s+)?References?\s*$",
-                which,
-                flags=re.IGNORECASE,
-            ):
-                part = "normative"
-            else:
-                part = "text"
-        parts[part] += line
-
-    refs: dict[str, list[str]] = {}
-    for part, content in parts.items():
-        refs[part] = re.findall(
-            r"(\[(?:\d+|[a-z]+(?:[-_.]?\w+)*)\]"
-            + (r"|RFC\d+|draft-[-a-z\d_.]+" if part == "text" else r"")
-            + r")",
-            unfold(content),
-            flags=re.IGNORECASE,
-        )
-        refs[part] = list({f"[{untag(ref)}]" for ref in refs[part]})
-
-    resolved: dict[str, list[str]] = {}
-    for part in ["informative", "normative"]:
-        resolved[part] = []
-        for ref in refs[part]:
-            ref_text = re.search(
-                r"\s*" + re.escape(ref) + r"\s+((?:[^\n][\n]?)+)\n",
-                parts[part],
-                re.DOTALL,
-            )
-            if ref_text:
-                ref_text = unfold(ref_text.group(0))
-                found = False
-
-                for pat in [r"(draft-[-a-z\d_.]+)", r"((?:RFC|rfc)\d+)"]:
-                    match = re.search(pat, ref_text)
-                    if match:
-                        found = True
-                        resolved[part].append((ref, match.group(0).lower()))
-                        break
-
-                if not found:
-                    urls = extract_urls(ref_text, True, True)
-                    resolved[part].append((ref, urls.pop() if urls else None))
-
-    resolved["text"] = refs["text"]
-    return resolved
-
-
 def untag(tag: str) -> str:
     """
     Remove angle brackets from reference tag.
@@ -303,53 +233,6 @@ def untag(tag: str) -> str:
     @return     Tag without angle brackets.
     """
     return re.sub(r"^\[(.*)\]$", r"\1", tag)
-
-
-def get_status(doc: str) -> str:
-    """
-    Extract the standards level status of a given document.
-
-    @param      doc   The document to extract the level from
-
-    @return     The status of the document.
-    """
-    status = re.search(
-        r"^(?:[Ii]ntended )?[Ss]tatus:\s*((?:\w+\s)+)",
-        doc,
-        re.MULTILINE,
-    )
-    return status.group(1).strip() if status else ""
-
-
-def get_relationships(
-    doc: str,
-) -> dict:
-    """
-    Extract the RFCs that are intended to be updated or obsoleted by this
-    document.
-
-    @param      doc   The document to extract the information from
-
-    @return     A dict of relationships and lists of RFC *numbers*
-    """
-    result = {}
-    pat = {"updates": r"[Uu]pdates", "obsoletes": r"[Oo]bsoletes"}
-    for rel in ["updates", "obsoletes"]:
-        match = re.search(
-            r"^"
-            + pat[rel]
-            + r":\s*((?:(?:RFC\s*)?\d{3,},?\s*)+)"
-            + r"(?:.*[\n\r\s]+((?:(?:RFC\s*)?\d{3,},?\s*)+)?)?",
-            doc,
-            re.MULTILINE,
-        )
-        if match:
-            result[rel] = "".join([group for group in match.groups() if group])
-            result[rel] = re.sub("rfc", "", result[rel], flags=re.IGNORECASE)
-            result[rel] = re.sub(r"[,\s]+(\w)", r",\1", result[rel])
-            result[rel] = result[rel].strip().split(",")
-            result[rel] = [r for r in result[rel] if r]
-    return result
 
 
 def basename(item: str) -> str:
